requirements.txt
PyOpenGL==3.1.9
PyQt6==6.8.1
numpy==2.2.3
opencv-python==4.11.0.86
pandas==2.2.3
pygame==2.6.1
